+++ packages/components/src/index.ts
@@ -17,6 +17,7 @@ export * from './components/form/bal-select/utils/bal-option.type'
 export * from './components/bal-tabs/bal-tab.type'
 export * from './components/form/bal-file-upload/bal-file-upload.type'
 export * from './components/notice/bal-modal/bal-modal.type'
+export * from './components/form/bal-radio/bal-radio.type'
 
 /**
  * Utils
@@ -28,7 +29,7 @@ export { isPlatform, getPlatforms, Platforms, PlatformSrcSet } from './utils/pla
 export { isBrowser, hasTouchSupport } from './utils/browser'
 export { scrollToFirstInvalidField } from './utils/form'
 export { BodyScrollBlocker } from './utils/toggle-scrolling-body'
-export { newBalRadioOption } from './components/form/bal-radio/bal-radio-group/bal-radio.util'
+export { newBalRadioOption } from './components/form/bal-radio/utils/bal-radio.util'
 
 /**
  * Controllers
